Try to fix warnings from TravisCI build
Most of these are switching from unsigned to size_t, since TravisCI's
build servers have size_t defined as unsigned long.

+++ src/Animation.cpp
@@ -361,7 +361,7 @@ unsigned short Animation::getLastFrameIndex(const short &frame) {
 	}
 	else {
 		// normal animation
-		for (unsigned i=frames.size(); i>0; i--) {
+		for (size_t i=frames.size(); i>0; i--) {
 			if (frames[i-1] == frame)
 				return static_cast<unsigned short>(i-1);
 		}
